requirements/base.txt
django==4.2.6
django-model-utils==4.3.1
django-redis==5.4.0
django-cors-headers==4.3.0

# un-versioned
slugify
GoogleAppEngineCloudStorageClient
black
requests
